+++ src/selection_indicator/style/rectangle.rs
@@ -1,5 +1,4 @@
 use embedded_graphics::{
-    pixelcolor::BinaryColor,
     prelude::DrawTarget,
     primitives::{Primitive, PrimitiveStyle, Rectangle as RectangleShape},
     Drawable,
@@ -15,21 +14,9 @@ use crate::{
 };
 
 #[derive(Clone, Copy)]
-pub struct Rectangle<T = BinaryColor> {
-    theme: T,
-}
+pub struct Rectangle;
 
-impl<T> Rectangle<T> {
-    pub fn new(theme: T) -> Self {
-        Self { theme }
-    }
-}
-
-impl<T> IndicatorStyle for Rectangle<T>
-where
-    T: Theme,
-{
-    type Theme = T;
+impl IndicatorStyle for Rectangle {
     type Shape = RectangleShape;
     type State = ();
 
@@ -46,18 +33,16 @@ where
         bounds
     }
 
-    fn color(&self, _state: &Self::State) -> <Self::Theme as Theme>::Color {
-        self.theme.selection_color()
-    }
-
-    fn draw<D>(
+    fn draw<T, D>(
         &self,
         state: &Self::State,
         input_state: InputState,
+        theme: &T,
         display: &mut D,
     ) -> Result<Self::Shape, D::Error>
     where
-        D: DrawTarget<Color = <Self::Theme as Theme>::Color>,
+        T: Theme,
+        D: DrawTarget<Color = T::Color>,
     {
         let display_area = display.bounding_box();
 
@@ -70,7 +55,7 @@ where
         let shape = self.shape(state, display_area, fill_width);
 
         shape
-            .into_styled(PrimitiveStyle::with_fill(self.theme.selection_color()))
+            .into_styled(PrimitiveStyle::with_fill(theme.selection_color()))
             .draw(display)?;
 
         Ok(shape)