+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Unreleased
  - Added initial color support
  - Style a menu by using a `Theme` object
  - Replaced the `Invert` selection indicator with `Rectangle`
+ - Removed `IndicatorStyle::Color`, `IndicatorStyle::color`, `IndicatorStyle::draw` is now generic over the theme
 
 0.5.4 (2023-10-27)
 ==================